Remove compact mode from buttons, move dropdown configuration
