+++ libs/core/src/lib/multi-input/multi-input-mobile/multi-input-mobile.component.html
@@ -28,7 +28,6 @@
                 <button fd-button
                         fdType="emphasized"
                         fd-dialog-decisive-button
-                        [compact]="true"
                         (click)="handleApprove()"
                 >
                     {{multiInputConfig?.approveButtonText}}
@@ -39,7 +38,6 @@
                 <button fd-button
                         fdType="transparent"
                         fd-dialog-decisive-button
-                        [compact]="true"
                         (click)="handleDismiss()"
                 >
                     {{multiInputConfig?.cancelButtonText}}
