+++ libs/core/src/lib/multi-input/multi-input-mobile/multi-input-mobile.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { MultiInputMobileComponent } from './multi-input-mobile.component';
-import { DropdownMobileConfiguration } from './dropdown-mobile-configuration';
+import { DropdownMobileConfiguration } from '../../utils/interfaces/dropdown-mobile-configuration';
 import { DynamicComponentService } from '../../utils/dynamic-component/dynamic-component.service';
 import { EventEmitter } from '@angular/core';
 import { DialogModule } from '../../dialog/dialog.module';
