+++ libs/core/src/lib/multi-input/public_api.ts
@@ -1,5 +1,5 @@
 export * from './multi-input.module';
 export * from './multi-input.component';
 export * from './multi-input-mobile/multi-input-mobile.module'
-export * from './multi-input-mobile/dropdown-mobile-configuration';
+export * from '../utils/interfaces/dropdown-mobile-configuration';
 export * from './multi-input-mobile/multi-input-mobile.component';
